refactor(test): split test_shared_labels to comply with LOC limits
Split TestCollectLabelAnomalies into separate file to reduce
test_shared_labels.py from 407 lines to 289 lines.

- Move TestCollectLabelAnomalies to test_collect_label_anomalies.py
- Both files now under 400 line limit
- All 68 tests still pass
